aws-cdk/python/splunk-indexer/requirements.txt
aws-cdk-lib==2.80.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-ram-alpha==2.80.0-alpha.0
boto3==1.28.0
